requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
feedparser==6.0.8
idna==3.3
oauthlib==3.1.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
sgmllib3k==1.0.0
tweepy==4.4.0
urllib3==1.26.7
